Rename vue component
